# pool-temp-backend/api-spec.yaml
openapi: 3.0.1
info:
  title: PoolTemp API V1
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /sensors/{sensorId}/measurements:
    post:
      tags:
        - measurements
      description: Create a new measurement
      parameters:
        - in: path
          name: sensorId
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        $ref: '#/components/requestBodies/MeasurementRequest'
      responses:
        201:
          $ref: '#/components/responses/MeasurementResponse'
        400:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'

    get:
      tags:
        - measurements
      description: Query existing measurements, ordered by timestamp descending, between start and end date
      parameters:
        - in: path
          name: sensorId
          schema:
            type: string
            format: uuid
          required: true
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
          required: false
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
          required: false
      responses:
        200:
          $ref: '#/components/responses/MeasurementListResponse'
        400:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'


  /sensors/{sensorId}/measurements/current:
    get:
      tags:
        - measurements
      description: Get the current temperature
      parameters:
        - in: path
          name: sensorId
          schema:
            type: string
            format: uuid
          required: true
      responses:
        200:
          $ref: '#/components/responses/MeasurementResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'

  /sensors:
    post:
      tags:
        - sensors
      description: Create a new sensor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sensor'
      responses:
        200:
          $ref: '#/components/responses/SensorResponse'
        400:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'

    get:
      tags:
        - sensors
      description: Get a list of all sensors
      responses:
        200:
          $ref: '#/components/responses/SensorListResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
components:
  requestBodies:
    MeasurementRequest:
      description: Represents a single measurement
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Measurement'

  responses:
    SensorListResponse:
      description: Represents a list of sensors
      content:
        application/json:
          schema:
            properties:
              items:
                type: array
                items:
                  $ref: '#/components/schemas/Sensor'

    SensorResponse:
      description: Represents a sensor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Sensor'

    MeasurementListResponse:
      description: Represents a single measurement
      content:
        application/json:
          schema:
            properties:
              items:
                type: array
                items:
                  $ref: '#/components/schemas/Measurement'

    MeasurementResponse:
      description: Represents a single measurement
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Measurement'

    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseBody'

  schemas:
    Sensor:
      required:
        - name
        - id
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          example: Pool Outdoor

    Measurement:
      required:
        - value
      properties:
        value:
          type: number
          format: decimal
          example: 20.3
        timestamp:
          type: string
          format: date-time
          example: 2017-07-21T17:32:28Z

    ErrorResponseBody:
      properties:
        statusCode:
          type: number
          example: 500
        message:
          type: string
          example: 'Internal server error'
        error:
          type: string
          example: 'Internal server error'